refactor: post,postapplicant 패키지 통합 의존성 리팩토링(#416)
