+++ backend/src/main/java/com/minibeit/post/domain/ApplyStatus.java
@@ -1,4 +1,4 @@
-package com.minibeit.postapplicant.domain;
+package com.minibeit.post.domain;
 
 public enum ApplyStatus {
     APPROVE, REJECT, WAIT, COMPLETE
