+++ backend/src/main/java/com/minibeit/post/domain/PostApplicant.java
@@ -1,8 +1,7 @@
-package com.minibeit.postapplicant.domain;
+package com.minibeit.post.domain;
 
 import com.minibeit.common.domain.BaseEntity;
 import com.minibeit.common.exception.PermissionException;
-import com.minibeit.post.domain.PostDoDate;
 import com.minibeit.user.domain.User;
 import lombok.*;
