+++ backend/src/main/java/com/minibeit/post/domain/PostValidator.java
@@ -1,9 +1,9 @@
-package com.minibeit.post.service;
+package com.minibeit.post.domain;
 
 import com.minibeit.businessprofile.domain.repository.UserBusinessProfileRepository;
 import com.minibeit.common.exception.InvalidOperationException;
 import com.minibeit.common.exception.PermissionException;
-import com.minibeit.postapplicant.domain.repository.PostApplicantRepository;
+import com.minibeit.post.domain.repository.PostApplicantRepository;
 import com.minibeit.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -13,11 +13,11 @@ import java.time.LocalDateTime;
 @Component
 @RequiredArgsConstructor
 public class PostValidator {
-    private final UserBusinessProfileRepository userBusinessProfileRepository;
-    private final PostApplicantRepository postApplicantRepository;
+        private final UserBusinessProfileRepository userBusinessProfileRepository;
+        private final PostApplicantRepository postApplicantRepository;
 
-    public void userInBusinessProfileValidate(Long businessProfileId, User user) {
-        if (!userBusinessProfileRepository.existsByUserIdAndBusinessProfileId(user.getId(), businessProfileId)) {
+        public void userInBusinessProfileValidate(Long businessProfileId, User user) {
+            if (!userBusinessProfileRepository.existsByUserIdAndBusinessProfileId(user.getId(), businessProfileId)) {
             throw new PermissionException("권한이 없습니다.");
         }
     }
