+++ backend/src/main/java/com/minibeit/post/domain/repository/PostFileRepository.java
@@ -1,6 +1,6 @@
-package com.minibeit.file.domain.repository;
+package com.minibeit.post.domain.repository;
 
-import com.minibeit.file.domain.PostFile;
+import com.minibeit.post.domain.PostFile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostFileRepository extends JpaRepository<PostFile, Long> {
