+++ backend/src/main/java/com/minibeit/post/domain/repository/PostRepositoryCustom.java
@@ -1,6 +1,6 @@
 package com.minibeit.post.domain.repository;
 
-import com.minibeit.postapplicant.domain.ApplyStatus;
+import com.minibeit.post.domain.ApplyStatus;
 import com.minibeit.post.domain.Payment;
 import com.minibeit.post.domain.Post;
 import com.minibeit.post.domain.PostStatus;
