+++ backend/src/main/java/com/minibeit/post/service/PostByBusinessService.java
@@ -4,23 +4,17 @@ import com.minibeit.businessprofile.domain.BusinessProfile;
 import com.minibeit.businessprofile.domain.repository.BusinessProfileRepository;
 import com.minibeit.businessprofile.service.exception.BusinessProfileNotFoundException;
 import com.minibeit.common.dto.PageDto;
-import com.minibeit.file.domain.PostFile;
-import com.minibeit.file.domain.repository.PostFileRepository;
+import com.minibeit.post.domain.*;
+import com.minibeit.post.domain.repository.PostFileRepository;
 import com.minibeit.file.service.S3Uploader;
 import com.minibeit.file.service.dto.SavedFile;
-import com.minibeit.post.domain.Post;
-import com.minibeit.post.domain.PostDoDate;
-import com.minibeit.post.domain.PostStatus;
-import com.minibeit.post.domain.RejectPost;
 import com.minibeit.post.domain.repository.PostDoDateRepository;
 import com.minibeit.post.domain.repository.PostRepository;
 import com.minibeit.post.domain.repository.RejectPostRepository;
 import com.minibeit.post.service.dto.PostRequest;
 import com.minibeit.post.service.dto.PostResponse;
 import com.minibeit.post.service.exception.PostNotFoundException;
-import com.minibeit.postapplicant.domain.ApplyStatus;
-import com.minibeit.postapplicant.domain.PostApplicant;
-import com.minibeit.postapplicant.domain.repository.PostApplicantRepository;
+import com.minibeit.post.domain.repository.PostApplicantRepository;
 import com.minibeit.school.domain.School;
 import com.minibeit.school.domain.SchoolRepository;
 import com.minibeit.user.domain.User;
