+++ backend/src/main/java/com/minibeit/post/web/PostController.java
@@ -2,7 +2,7 @@ package com.minibeit.post.web;
 
 import com.minibeit.common.dto.ApiResult;
 import com.minibeit.common.dto.PageDto;
-import com.minibeit.postapplicant.domain.ApplyStatus;
+import com.minibeit.post.domain.ApplyStatus;
 import com.minibeit.post.domain.Payment;
 import com.minibeit.post.service.dto.PostResponse;
 import com.minibeit.post.service.PostService;
