+++ backend/src/main/java/com/minibeit/review/service/BusinessUserReviewService.java
@@ -3,13 +3,10 @@ package com.minibeit.review.service;
 import com.minibeit.businessprofile.domain.BusinessProfile;
 import com.minibeit.businessprofile.domain.repository.BusinessProfileRepository;
 import com.minibeit.businessprofile.service.exception.BusinessProfileNotFoundException;
-import com.minibeit.postapplicant.domain.PostApplicant;
-import com.minibeit.postapplicant.domain.repository.PostApplicantRepository;
-import com.minibeit.postapplicant.service.exception.PostApplicantNotFoundException;
-import com.minibeit.review.domain.BusinessUserReview;
-import com.minibeit.review.domain.BusinessUserReviewDetail;
-import com.minibeit.review.domain.BusinessUserReviewEvalType;
-import com.minibeit.review.domain.BusinessUserReviewType;
+import com.minibeit.post.domain.PostApplicant;
+import com.minibeit.post.domain.repository.PostApplicantRepository;
+import com.minibeit.post.service.exception.PostApplicantNotFoundException;
+import com.minibeit.review.domain.*;
 import com.minibeit.review.domain.repository.BusinessUserReviewDetailRepository;
 import com.minibeit.review.domain.repository.BusinessUserReviewRepository;
 import com.minibeit.review.service.dto.BusinessUserReviewResponse;
