+++ backend/src/test/java/com/minibeit/businessprofile/web/BusinessProfileControllerTest.java
@@ -9,7 +9,7 @@ import com.minibeit.businessprofile.service.BusinessProfileService;
 import com.minibeit.post.domain.Payment;
 import com.minibeit.post.domain.Post;
 import com.minibeit.post.domain.PostDoDate;
-import com.minibeit.file.domain.PostFile;
+import com.minibeit.post.domain.PostFile;
 import com.minibeit.school.domain.School;
 import com.minibeit.user.domain.Gender;
 import com.minibeit.user.domain.Role;
