+++ backend/src/test/java/com/minibeit/post/service/PostApplicantByBusinessFindServiceTest.java
@@ -1,4 +1,4 @@
-package com.minibeit.postapplicant.service;
+package com.minibeit.post.service;
 
 import com.minibeit.ServiceIntegrationTest;
 import com.minibeit.businessprofile.domain.BusinessProfile;
@@ -6,15 +6,16 @@ import com.minibeit.businessprofile.domain.UserBusinessProfile;
 import com.minibeit.businessprofile.domain.repository.BusinessProfileRepository;
 import com.minibeit.businessprofile.domain.repository.UserBusinessProfileRepository;
 import com.minibeit.post.domain.*;
-import com.minibeit.postapplicant.domain.repository.PostApplicantRepository;
+import com.minibeit.post.domain.repository.PostApplicantRepository;
 import com.minibeit.post.domain.repository.PostDoDateRepository;
 import com.minibeit.post.domain.repository.PostRepository;
-import com.minibeit.postapplicant.service.dto.PostApplicantDto;
-import com.minibeit.postapplicant.service.dto.PostApplicantResponse;
+import com.minibeit.post.service.PostApplicantByBusinessService;
+import com.minibeit.post.service.dto.PostApplicantDto;
+import com.minibeit.post.service.dto.PostApplicantResponse;
 import com.minibeit.post.service.dto.PostDto;
 import com.minibeit.post.service.dto.PostRequest;
-import com.minibeit.postapplicant.domain.ApplyStatus;
-import com.minibeit.postapplicant.domain.PostApplicant;
+import com.minibeit.post.domain.ApplyStatus;
+import com.minibeit.post.domain.PostApplicant;
 import com.minibeit.school.domain.School;
 import com.minibeit.school.domain.SchoolRepository;
 import com.minibeit.user.domain.Role;
