+++ backend/src/test/java/com/minibeit/post/service/PostApplicantByBusinessServiceTest.java
@@ -1,4 +1,4 @@
-package com.minibeit.postapplicant.service;
+package com.minibeit.post.service;
 
 import com.minibeit.ServiceIntegrationTest;
 import com.minibeit.businessprofile.domain.BusinessProfile;
@@ -14,13 +14,14 @@ import com.minibeit.post.domain.RejectPost;
 import com.minibeit.post.domain.repository.PostDoDateRepository;
 import com.minibeit.post.domain.repository.PostRepository;
 import com.minibeit.post.domain.repository.RejectPostRepository;
+import com.minibeit.post.service.PostApplicantByBusinessService;
 import com.minibeit.post.service.dto.PostDto;
 import com.minibeit.post.service.dto.PostRequest;
-import com.minibeit.postapplicant.domain.ApplyStatus;
-import com.minibeit.postapplicant.domain.PostApplicant;
-import com.minibeit.postapplicant.domain.repository.PostApplicantRepository;
-import com.minibeit.postapplicant.service.dto.PostApplicantRequest;
-import com.minibeit.postapplicant.service.exception.PostApplicantNotFoundException;
+import com.minibeit.post.domain.ApplyStatus;
+import com.minibeit.post.domain.PostApplicant;
+import com.minibeit.post.domain.repository.PostApplicantRepository;
+import com.minibeit.post.service.dto.PostApplicantRequest;
+import com.minibeit.post.service.exception.PostApplicantNotFoundException;
 import com.minibeit.school.domain.School;
 import com.minibeit.school.domain.SchoolRepository;
 import com.minibeit.user.domain.Role;
