+++ backend/src/test/java/com/minibeit/post/service/PostByBusinessServiceTest.java
@@ -9,8 +9,8 @@ import com.minibeit.common.exception.InvalidOperationException;
 import com.minibeit.common.exception.PermissionException;
 import com.minibeit.file.domain.FileServer;
 import com.minibeit.file.domain.FileType;
-import com.minibeit.file.domain.PostFile;
-import com.minibeit.file.domain.repository.PostFileRepository;
+import com.minibeit.post.domain.PostFile;
+import com.minibeit.post.domain.repository.PostFileRepository;
 import com.minibeit.file.service.S3Uploader;
 import com.minibeit.file.service.dto.SavedFile;
 import com.minibeit.post.domain.*;
@@ -22,10 +22,10 @@ import com.minibeit.post.service.dto.PostDto;
 import com.minibeit.post.service.dto.PostRequest;
 import com.minibeit.post.service.dto.PostResponse;
 import com.minibeit.post.service.exception.PostNotFoundException;
-import com.minibeit.postapplicant.domain.ApplyStatus;
-import com.minibeit.postapplicant.domain.PostApplicant;
-import com.minibeit.postapplicant.domain.repository.PostApplicantRepository;
-import com.minibeit.postapplicant.service.exception.PostApplicantNotFoundException;
+import com.minibeit.post.domain.ApplyStatus;
+import com.minibeit.post.domain.PostApplicant;
+import com.minibeit.post.domain.repository.PostApplicantRepository;
+import com.minibeit.post.service.exception.PostApplicantNotFoundException;
 import com.minibeit.school.domain.School;
 import com.minibeit.school.domain.SchoolRepository;
 import com.minibeit.user.domain.Role;
