+++ backend/src/test/java/com/minibeit/post/service/PostFindServiceTest.java
@@ -8,8 +8,8 @@ import com.minibeit.businessprofile.domain.repository.UserBusinessProfileReposit
 import com.minibeit.common.dto.PageDto;
 import com.minibeit.file.domain.FileServer;
 import com.minibeit.file.domain.FileType;
-import com.minibeit.file.domain.PostFile;
-import com.minibeit.file.domain.repository.PostFileRepository;
+import com.minibeit.post.domain.PostFile;
+import com.minibeit.post.domain.repository.PostFileRepository;
 import com.minibeit.file.service.dto.SavedFile;
 import com.minibeit.post.domain.Payment;
 import com.minibeit.post.domain.Post;
@@ -22,9 +22,9 @@ import com.minibeit.post.service.dto.PostDto;
 import com.minibeit.post.service.dto.PostRequest;
 import com.minibeit.post.service.dto.PostResponse;
 import com.minibeit.post.service.exception.PostNotFoundException;
-import com.minibeit.postapplicant.domain.ApplyStatus;
-import com.minibeit.postapplicant.domain.PostApplicant;
-import com.minibeit.postapplicant.domain.repository.PostApplicantRepository;
+import com.minibeit.post.domain.ApplyStatus;
+import com.minibeit.post.domain.PostApplicant;
+import com.minibeit.post.domain.repository.PostApplicantRepository;
 import com.minibeit.school.domain.School;
 import com.minibeit.school.domain.SchoolRepository;
 import com.minibeit.security.userdetails.CustomUserDetails;
