+++ backend/src/test/java/com/minibeit/post/web/PostApplicantControllerTest.java
@@ -1,7 +1,8 @@
-package com.minibeit.postapplicant.web;
+package com.minibeit.post.web;
 
 import com.minibeit.MvcTest;
-import com.minibeit.postapplicant.service.PostApplicantService;
+import com.minibeit.post.service.PostApplicantService;
+import com.minibeit.post.web.PostApplicantController;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
