+++ backend/src/test/java/com/minibeit/post/web/PostControllerTest.java
@@ -3,11 +3,11 @@ package com.minibeit.post.web;
 import com.minibeit.MvcTest;
 import com.minibeit.file.domain.Avatar;
 import com.minibeit.businessprofile.domain.BusinessProfile;
-import com.minibeit.file.domain.PostFile;
+import com.minibeit.post.domain.PostFile;
 import com.minibeit.post.domain.*;
 import com.minibeit.post.service.dto.PostResponse;
 import com.minibeit.post.service.PostService;
-import com.minibeit.postapplicant.domain.ApplyStatus;
+import com.minibeit.post.domain.ApplyStatus;
 import com.minibeit.school.domain.School;
 import com.minibeit.security.userdetails.CustomUserDetails;
 import com.minibeit.user.domain.User;
